src/requirements.txt
pydantic>=2.5
httpx>=0.27
fastapi>=0.109
uvicorn>=0.27
jinja2>=3.1
python-dotenv>=1.0
pytest>=8.0
pytest-asyncio>=0.23
